clean up lib.rs
