+++ src/buffer.rs
@@ -1,3 +1,5 @@
+#![cfg(not(Py_LIMITED_API))]
+#![cfg_attr(docsrs, doc(cfg(not(Py_LIMITED_API))))]
 // Copyright (c) 2017 Daniel Grunwald
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy of this
