+++ src/conversions/mod.rs
@@ -7,3 +7,4 @@ pub mod num_bigint;
 pub mod num_complex;
 mod osstr;
 mod path;
+pub mod serde;
