+++ src/conversions/serde.rs
@@ -1,3 +1,23 @@
+#![cfg_attr(docsrs, doc(cfg(feature = "serde")))]
+#![cfg(feature = "serde")]
+
+//! Enables (de)serialization of [`Py`]`<T>` objects via [serde](https://docs.rs/serde).
+//!
+//! # Setup
+//!
+//! To use this feature, add this to your **`Cargo.toml`**:
+//!
+//! ```toml
+//! [dependencies]
+//! serde = "1.0"
+// workaround for `extended_key_value_attributes`: https://github.com/rust-lang/rust/issues/82768#issuecomment-803935643
+#![cfg_attr(docsrs, cfg_attr(docsrs, doc = concat!("pyo3 = { version = \"", env!("CARGO_PKG_VERSION"),  "\", features = [\"serde\"] }")))]
+#![cfg_attr(
+    not(docsrs),
+    doc = "pyo3 = { version = \"*\", features = [\"serde\"] }"
+)]
+//! ```
+
 use crate::{Py, PyAny, PyClass, Python};
 use serde::{de, ser, Deserialize, Deserializer, Serialize, Serializer};
