+++ src/impl_.rs
@@ -1,6 +1,8 @@
-//! Internals of PyO3 which are accessed by code expanded from PyO3's procedural macros. Usage of
-//! any of these APIs in downstream code is implicitly acknowledging that these APIs may change at
-//! any time without documentation in the CHANGELOG and without breaking semver guarantees.
+//! Internals of PyO3 which are accessed by code expanded from PyO3's procedural macros.
+//!
+//! Usage of any of these APIs in downstream code is implicitly acknowledging that these
+//! APIs may may change at any time without documentation in the CHANGELOG and without
+//! breaking semver guarantees.
 
 pub mod deprecations;
 pub mod freelist;
