+++ src/prelude.rs
@@ -1,8 +1,8 @@
 // Copyright (c) 2017-present PyO3 Project and Contributors
 
-//! A collection of items you most likely want to have in scope when working with pyo3
+//! PyO3's prelude.
 //!
-//! The purpose of this module is to alleviate imports of many common pyo3 traits
+//! The purpose of this module is to alleviate imports of many commonly used items of the PyO3 crate
 //! by adding a glob import to the top of pyo3 heavy modules:
 //!
 //! ```
@@ -16,9 +16,10 @@ pub use crate::instance::{Py, PyObject};
 pub use crate::pycell::{PyCell, PyRef, PyRefMut};
 pub use crate::pyclass_init::PyClassInitializer;
 pub use crate::python::Python;
-pub use crate::{FromPyObject, IntoPy, IntoPyPointer, PyTryFrom, PyTryInto, ToPyObject};
-// PyModule is only part of the prelude because we need it for the pymodule function
 pub use crate::types::{PyAny, PyModule};
 pub use crate::wrap_pyfunction;
+pub use crate::{FromPyObject, IntoPy, IntoPyPointer, PyTryFrom, PyTryInto, ToPyObject};
+
+#[cfg_attr(docsrs, doc(cfg(feature = "macros")))]
 #[cfg(feature = "macros")]
-pub use {crate::proc_macro::*, pyo3_macros::FromPyObject};
+pub use pyo3_macros::{pyclass, pyfunction, pymethods, pymodule, pyproto, FromPyObject};
